Changed ferry location coordinates format to float, fixed price for route-not-found case, fixed switching between two-way and one-way routes when defining same location for start and end, sorting order data via new "order" key in order form values before sending to server and when adding new location controls, fixed-lines-appearing-between-unused-locations bug in map-controls component

+++ models/ferries_directions.js
@@ -2,8 +2,8 @@
 module.exports = (sequelize, DataTypes) => {
   const ferries_directions = sequelize.define('ferries_directions', {
     name: DataTypes.STRING,
-    latitude: DataTypes.DECIMAL(10, 6),
-    longitude: DataTypes.DECIMAL(11, 6)
+    latitude: DataTypes.FLOAT,
+    longitude: DataTypes.FLOAT
   }, {underscored: true});
   ferries_directions.associate = function(models) {
     // associations can be defined here
